Initial commit.
"If you know the enemy and know yourself, you need not fear
the result of a hundred battles. If you know yourself but not the
enemy, for every victory gained you will also suffer a defeat. If
you know neither the enemy nor yourself, you will succumb in
every battle."
-- Sun Tzu, The Art Of War